LPS-72545 Move to the proper packages
