+++ modules/apps/expando/expando-export-import/src/main/java/com/liferay/expando/exportimport/internal/model/adapter/builder/StagedExpandoColumnModelAdapterBuilder.java
@@ -12,10 +12,11 @@
  * details.
  */
 
-package com.liferay.expando.exportimport.internal.model.adapter;
+package com.liferay.expando.exportimport.internal.model.adapter.builder;
 
-import com.liferay.expando.kernel.model.ExpandoTable;
-import com.liferay.expando.kernel.model.adapter.StagedExpandoTable;
+import com.liferay.expando.exportimport.internal.model.adapter.StagedExpandoColumnImpl;
+import com.liferay.expando.kernel.model.ExpandoColumn;
+import com.liferay.expando.kernel.model.adapter.StagedExpandoColumn;
 import com.liferay.portal.kernel.model.adapter.builder.ModelAdapterBuilder;
 
 import org.osgi.service.component.annotations.Component;
@@ -24,12 +25,12 @@ import org.osgi.service.component.annotations.Component;
  * @author Akos Thurzo
  */
 @Component(immediate = true, service = ModelAdapterBuilder.class)
-public class StagedExpandoTableModelAdapterBuilder
-	implements ModelAdapterBuilder<ExpandoTable, StagedExpandoTable> {
+public class StagedExpandoColumnModelAdapterBuilder
+	implements ModelAdapterBuilder<ExpandoColumn, StagedExpandoColumn> {
 
 	@Override
-	public StagedExpandoTable build(ExpandoTable expandoTable) {
-		return new StagedExpandoTableImpl(expandoTable);
+	public StagedExpandoColumn build(ExpandoColumn expandoColumn) {
+		return new StagedExpandoColumnImpl(expandoColumn);
 	}
 
 }
